orders/requirements.txt
Django
djangorestframework
djongo
psycopg2
gunicorn
requests
django-rest-swagger